レビューとトップページ修正

+++ sources/web/database/migrations/2018_05_16_000000_create_review_disable_terms.php
@@ -18,7 +18,7 @@ class CreateReviewDisableTerms extends Migration
     {
         Schema::create('review_disable_terms', function (Blueprint $table) {
             $table->unsignedInteger('user_id')->comment('ユーザーID');
-            $table->unsignedTinyInteger('soft_id')->comment('ソフトID');
+            $table->unsignedInteger('soft_id')->comment('ソフトID');
             $table->date('start_date')->comment('無効化開始日時');
             $table->timestamps();
             $table->primary(['user_id', 'soft_id']);
